option to add custom addresses and account nav in separate file

+++ OnlineStore/src/main/java/com/store/models/Address.java
@@ -10,20 +10,20 @@ public class Address
 {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String label;
 	private String country;
 	private String city;
 	private String street;
 	private String zip;
-	private int userId;
+	private Long userId;
 	
 	public Address()
 	{
 		
 	}
 
-	public Address(int id, String label, String country, String city, String street, String zip, int userId)
+	public Address(Long id, String label, String country, String city, String street, String zip, Long userId)
 	{
 		super();
 		this.id = id;
@@ -34,12 +34,12 @@ public class Address
 		this.userId = userId;
 	}
 
-	public int getId()
+	public Long getId()
 	{
 		return id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
@@ -94,12 +94,12 @@ public class Address
 		this.zip = zip;
 	}
 
-	public int getUserId()
+	public Long getUserId()
 	{
 		return userId;
 	}
 
-	public void setUserId(int userId)
+	public void setUserId(Long userId)
 	{
 		this.userId = userId;
 	}
